move forget pebble class and rename
